Interface to berkeleydb package added.
The bsddb3 package is deprecated and not supported after Python3.9 but
is still needed at Python3.5 and earlier, so the inteface to bsddb3
is not removed.  The berkeleydb package is preferred over the bsddb3
package when creating new databases, and when accessing existing
Berkeley DB databases of chess results, if both are installed.