Add ability to upload results submissions to https://www.ecfrating.org.uk/ using the API defined at ecfrating.org.uk/v2/submit/help.php and save responses in a form suitable for updating ECF codes on local database.
HTTPS POSTs are done with the curl utility if available, falling back to
the requests package if installed.

The email responses currently do not have the information needed to do the
local database updates.  It was, and remains, possible to do the updates
manually for each ECF code.

It is still necessary to use the browser interface to request new ECF codes
before actually allocating them in a submission using the facilities added
here.

All dates in the ECF responses, in the expected yyyy-mm-dd and dd/mm/yyyy
formats, are redacted to 'nnnn-nn-nn' and 'nn/nn/nnnn' respectively because
some of them could be dates of birth.  None of the dates are relevant to
updating ECF codes.  The responses are saved to file before applying the
updates from file to local database.